+++ ICSharpCode.Decompiler/TypeSystem/UsingScope.cs
@@ -23,12 +23,12 @@ using System.Collections.Immutable;
 
 using ICSharpCode.Decompiler.CSharp.Syntax;
 using ICSharpCode.Decompiler.Semantics;
-using ICSharpCode.Decompiler.TypeSystem;
+using ICSharpCode.Decompiler.TypeSystem.Implementation;
 using ICSharpCode.Decompiler.Util;
 
 #nullable enable
 
-namespace ICSharpCode.Decompiler.CSharp.TypeSystem
+namespace ICSharpCode.Decompiler.TypeSystem
 {
 	/// <summary>
 	/// Represents a scope that contains "using" statements.
